Plot time-steps instead of time-bins
