+++ colibre/config.yml
@@ -262,11 +262,11 @@ scripts:
     title: Wall-clock time per time-bin
     section: Run Performance
     output_file: wallclock_timebin_hist.png
-  - filename: scripts/bh_timebin_hist.py
-    caption: Distributions of black-hole time-steps (shown as time-bins). The solid line-styled histogram shows BH current time-steps. The dashed line-styled histogram shows BH minimal time-steps reached during the simulation. The dotted line-styled histogram shows accretion-limited time-steps of BHs. The dash-dotted and dashed vertical lines indicate, respectively, the BH minimum allowed time-bin in the simulation and the age of the Universe.
-    title: Black-hole time-bins histogram
+  - filename: scripts/bh_timestep_hist.py
+    caption: Distributions of black-hole time-steps. The solid curve shows BH current time-steps. The dashed curve shows BH minimal time-steps reached during the simulation. The dotted curve shows accretion-limited time-steps of BHs. The dash-dotted and dashed vertical lines indicate, respectively, the BH minimum allowed time-step based on accretion rate and the age of the Universe.
+    title: Black-hole time-steps histogram
     section: Run Performance
-    output_file: bh_timebin_hist.png
+    output_file: bh_timestep_hist.png
   - filename: scripts/deadtime_timebin_hist.py
     caption: Dead time fraction in each simulation time-bin.
     title: Deadtime fraction per time-bin
